create relations subscription and team with campaign

+++ src/main/java/br/com/hpaiva/campaignservice/campaign/CampaignRequest.java
@@ -19,6 +19,6 @@ public class CampaignRequest {
 
     private LocalDateTime endEffectiveDate;
 
-    private Integer idHeartTeam;
+    private Long idHeartTeam;
 
 }
